qnd: don't delete the loginData if no loginData found
git-svn-id: svn+ssh://kelev.kaltura.com/usr/local/kalsource/backend/server/trunk/core@88990 6b8eccd3-e8c5-4e7d-8186-e12b5326b719

+++ alpha/lib/model/UserLoginDataPeer.php
@@ -609,7 +609,8 @@ class UserLoginDataPeer extends BaseUserLoginDataPeer {
 		
 		if ($countUsers <= 0) {
 			$loginData = self::retrieveByPK($loginDataId);
-			$loginData->delete();
+			if($loginData)
+				$loginData->delete();
 		}
